fix.backend: save to repo only if the image list isn't empty
This commit solves the bug which occurs when a merchant wants to
update her product. If she doesn't upload images, it would remove
the previous images.

+++ src/main/java/com/fosskeeters/rrss/controllers/ProductController.java
@@ -250,8 +250,10 @@ public class ProductController {
                 displayOrder++;
             }
 
-            productImageRepository.deleteAllByProduct(product.get());
-            productImageRepository.saveAll(productImages);
+            if (!productImages.isEmpty()) {
+                productImageRepository.deleteAllByProduct(product.get());
+                productImageRepository.saveAll(productImages);
+            }
         }
 
         product.get().setFromProductDto(productDto);
